+++ pymcfunc/__init__.py
@@ -4,4 +4,4 @@ from pymcfunc.pack import *
 import pymcfunc.selectors as sel
 from pymcfunc.coords import *
 
-__version__ = "0.0"
+__version__ = "0.2"