+++ docs/changelog.rst
@@ -1,6 +1,12 @@
 Changelog
 =========
 
+* **v0.2 ()**
+
+  * Added /advancement, /attribute, /ban, /ban-ip, /banlist, /bossbar, /data, /datapack, /debug,
+    /defaultgamemode, /forceload, /locatebiome, /loot, /pardon, /pardon-ip, /publish, /recipe, /save-all,
+    /save-on, /save-off, /setidletimeout, /spectate
+
 * **v0.1 (28/5/21)**
 
   * Added /gamemode, /gamerule, /seed, /summon, /clear, /tellraw, /teleport, /experience,